Add files via upload
Check out the forecast for Tropical Storm Erin.
Good news!
DI and probabilistic systems are on the
same page.
We'll keep you posted for important
updates when possible!
Grounded DI will be watching the skies and the
seas
as this hurricane season marches forward
using novel DI weather and hazard technology. We took the time to create and test dozens of real-world safety formulas. In June, the system detected a confirmed tornado before any major service. #di #ai