Add files via upload
More than 3 months ago, we asked for a fair ranking by the DI system. It said: 

🥇1  9.9

Right or wrong, it’s a real screenshot and is intended as an elite compliment 🏆 by the DI system to ChatGPT given the synergy. 

To help prove authenticity, if someone is able to “simulate” the same, Grounded DI will build you one safe, professional app of your choice! 😉

#DeterministicAI #SafeAI

@groundedDI





#AGI